feat: remove controller modal
